Add FactoryBot/AttributeDefinedStatically Cop
We are planning to deprecate static attributes in
[`factory_bot#1135`][pull], and it would be nice to have a simple way to
autocorrect all static attributes into dynamic attributes.

This new cop is not compatible with the existing
`FactoryBot/DynamicAttributeDefinedStatically` and
`FactoryBot/StaticAttributeDefinedDynamically` cops, so we disabled
those in `config/default.yml`.

It probably makes sense to deprecate those in a future pull request, or
keep them and have them implemented as no-ops.

Co-authored-by: Sean Doyle <sean.p.doyle24@gmail.com>

[pull]: https://github.com/thoughtbot/factory_bot/pull/1135